apps/agent-service/requirements.txt
# FastAPI Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.5.0

# Database - MySQL (FIXED: was MongoDB)
mysql-connector-python>=8.2.0
# Alternative: pymysql>=1.1.0

# LangChain
langchain>=0.3.0
langchain-community>=0.3.0
langchain-ollama>=0.3.0

# Embeddings (Local - FREE)
sentence-transformers>=2.3.0

# Vector Store (Local - FREE)
chromadb>=0.4.0

# Web Search
tavily-python>=0.3.0

# Utilities
python-multipart>=0.0.6
requests>=2.31.0
numpy>=1.24.0

# Optional: For better async support
aiohttp>=3.9.0
